pyproject.toml
[project]
name = "modal-fish-s2-pro"
version = "0.2.0"
description = "Real-time voice AI agent: Pipecat + Whisper + Smart Turn + OpenRouter + Fish S2 Pro"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[whisper,openrouter,fish,daily,silero,local-smart-turn]>=0.0.70",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]
